+++ examples/python/reset.py
@@ -30,7 +30,7 @@ if __name__ == '__main__':
 
     # (assert (= x #b010))
     bitwuzla.assert_formula(
-        mk_term(Kind.EQUAL, [x, mk_bv_value(sortbv3, 2, 10)]))
+        mk_term(Kind.EQUAL, [x, mk_bv_value(sortbv3, 2)]))
     # (check-sat)
     result = bitwuzla.check_sat()
     print('Expect: sat')
@@ -51,11 +51,11 @@ if __name__ == '__main__':
 
     # (assert (= x #b011))
     bitwuzla.assert_formula(
-        mk_term(Kind.EQUAL, [x, mk_bv_value(sortbv3, 3, 10)]))
+        mk_term(Kind.EQUAL, [x, mk_bv_value(sortbv3, 3)]))
     # (assert (= x (select a #b01)))
     bitwuzla.assert_formula(mk_term(
         Kind.EQUAL,
-        [x, mk_term(Kind.ARRAY_SELECT, [a, mk_bv_value(sortbv2, 1, 10)])]))
+        [x, mk_term(Kind.ARRAY_SELECT, [a, mk_bv_value(sortbv2, 1)])]))
     # (check-sat)
     result = bitwuzla.check_sat()
     print('Expect: sat')